Move and split api bulk delete tests
